Stripping undefined properties from entities on conversion
Undefined values shouldn't continually exist on objects but they will still show up under Object.keys which is significant if things are iterating through them. This is a utility function that may need to be hoisted to @uems/micro-builder to ensure that its available across all services but it will remain here for now. Types are coerced because this modification should not invalidate a type because undefined properties don't need to exist.